Fix rebase conflicts

+++ core/src/main/scala/latis/dsl/package.scala
@@ -50,7 +50,7 @@ package object dsl {
     def first(): Dataset                           = dataset.withOperation(Head())
     def last(): Dataset                            = dataset.withOperation(Last())
     def tail(): Dataset                            = dataset.withOperation(Tail())
-    def take(n: Long): Dataset                     = dataset.withOperation(Take(n))
+    def take(n: Int): Dataset                      = dataset.withOperation(Take(n))
     def takeRight(n: Int): Dataset                 = dataset.withOperation(TakeRight(n))
     def transpose(): Dataset                       = dataset.withOperation(Transpose())
